2025/requirements.txt
requests
python-dotenv
av
pillow
numpy
scipy
